WIP: parse functions now accept tags.
