+++ src/test/scala/net/jcazevedo/moultingyaml/RoundTripSpec.scala
@@ -2,12 +2,8 @@ package net.jcazevedo.moultingyaml
 
 import java.io.File
 import java.net.URLDecoder
-
-import scala.io.Source
-
 import org.specs2.mutable.Specification
-
-import net.jcazevedo.moultingyaml.defaultParser._
+import scala.io.Source
 
 class RoundTripSpec extends Specification {
   def getResourceURL(resource: String): String =
@@ -17,9 +13,9 @@ class RoundTripSpec extends Specification {
     "work in a round trip fashion" in {
       val files = new File(getResourceURL("/examples")).listFiles()
       forall(files) { file =>
-        val yamls = Source.fromFile(file).mkString.parseYamls
+        val yamls = Source.fromFile(file).mkString.parseYamls()
         forall(yamls) { innerYaml =>
-          innerYaml.prettyPrint.parseYaml mustEqual innerYaml
+          innerYaml.prettyPrint.parseYaml() mustEqual innerYaml
         }
       }
     }
