+++ src/test/scala/net/jcazevedo/moultingyaml/YamlParserSpec.scala
@@ -1,15 +1,11 @@
 package net.jcazevedo.moultingyaml
 
-import java.net.URLDecoder
-
-import scala.io.Source
-import scala.util.{ Failure, Success, Try }
-
 import com.github.nscala_time.time.Imports._
+import java.net.URLDecoder
 import org.specs2.execute.Result
 import org.specs2.mutable._
-
-import net.jcazevedo.moultingyaml.defaultParser._
+import scala.io.Source
+import scala.util.{ Failure, Success, Try }
 
 class YamlParserSpec extends Specification {
   def getResourceURL(resource: String): String =
@@ -20,7 +16,7 @@ class YamlParserSpec extends Specification {
   }
 
   def withYamls(filename: String)(f: Seq[YamlValue] => Result): Result = {
-    Try(Source.fromFile(getResourceURL(filename)).mkString.parseYamls) match {
+    Try(Source.fromFile(getResourceURL(filename)).mkString.parseYamls()) match {
       case Success(yamls) => f(yamls)
       case Failure(error) =>
         error.printStackTrace()
